+++ src/components/badge/badge.stories.tsx
@@ -1,12 +1,51 @@
 import React from 'react';
-import { Badge } from '.';
+import { Badge, BadgeProps } from '.';
+import { Story, Meta } from '@storybook/react/types-6-0';
+import { Box, Flex } from '../layout';
 
 export default {
-  title: 'Badge',
+  title: 'Components/Badge',
+  component: Badge,
+  argTypes: {
+    backgroundColor: { control: 'color' },
+    color: { control: 'color' },
+  },
+} as Meta;
+
+const Template: Story<BadgeProps> = (args) => <Badge {...args} />;
+
+// Base default button
+export const Base = Template.bind({});
+Base.args = {
+  text: 'Say Hello',
 };
 
-export const Primary = () => (
-  <Badge size='default' backgroundColor='pink'>
-    M
-  </Badge>
+// sizes
+export const Sizes = (args: BadgeProps) => (
+  <Flex>
+    <Badge {...args} size='large' text='Preview' />
+    <Badge {...args} size='medium' text='Preview' />
+    <Badge {...args} size='small' text='Preview' />
+  </Flex>
+);
+
+// shape
+export const Shape = (args: BadgeProps) => (
+  <Box>
+    <Flex>
+      <Badge {...args} shape='square' size='large' text='Preview' />
+      <Badge {...args} shape='square' size='medium' text='Preview' />
+      <Badge {...args} shape='square' size='small' text='Preview' />
+    </Flex>
+    <Flex>
+      <Badge {...args} shape='pill' size='large' text='Preview' />
+      <Badge {...args} shape='pill' size='medium' text='Preview' />
+      <Badge {...args} shape='pill' size='small' text='Preview' />
+    </Flex>
+    <Flex>
+      <Badge {...args} shape='circle' size='large' text='M' />
+      <Badge {...args} shape='circle' size='medium' text='M' />
+      <Badge {...args} shape='circle' size='small' text='M' />
+    </Flex>
+  </Box>
 );